feat(desktop): add RDF/XML import support with format adapter architecture
Introduce a pluggable FormatAdapter registry that resolves parsers by file
extension. Extract shared quad-walking logic into quads.ts so both Turtle
and RDF/XML parsers produce identical Ontology structures. Wire .rdf/.owl
extensions through the open-file dialog and store.
