+++ apps/desktop/package.json
@@ -38,6 +38,7 @@
     "@radix-ui/react-slot": "^1.2.4",
     "@radix-ui/react-tabs": "^1.1.13",
     "@radix-ui/react-tooltip": "^1.2.8",
+    "@rdfjs/types": "^2.0.1",
     "@sentry/electron": "^7.10.0",
     "@streamdown/code": "^1.1.1",
     "@tailwindcss/vite": "4",
@@ -51,6 +52,7 @@
     "motion": "^12.38.0",
     "n3": "^2.0.3",
     "posthog-js": "^1.364.1",
+    "rdfxml-streaming-parser": "^3.2.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-resizable-panels": "^4.7.6",
