+++ apps/desktop/resources/sample-ontologies/people.rdf
@@ -1,0 +1,73 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<rdf:RDF
+    xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
+    xmlns:rdfs="http://www.w3.org/2000/01/rdf-schema#"
+    xmlns:owl="http://www.w3.org/2002/07/owl#"
+    xmlns:xsd="http://www.w3.org/2001/XMLSchema#"
+    xmlns:ex="http://example.org/ontology#">
+
+    <owl:Class rdf:about="http://example.org/ontology#Person">
+        <rdfs:label>Person</rdfs:label>
+        <rdfs:comment>A human being</rdfs:comment>
+    </owl:Class>
+
+    <owl:Class rdf:about="http://example.org/ontology#Organisation">
+        <rdfs:label>Organisation</rdfs:label>
+        <rdfs:comment>A formal group of people</rdfs:comment>
+    </owl:Class>
+
+    <owl:Class rdf:about="http://example.org/ontology#Employee">
+        <rdfs:label>Employee</rdfs:label>
+        <rdfs:subClassOf rdf:resource="http://example.org/ontology#Person"/>
+    </owl:Class>
+
+    <owl:Class rdf:about="http://example.org/ontology#Manager">
+        <rdfs:label>Manager</rdfs:label>
+        <rdfs:subClassOf rdf:resource="http://example.org/ontology#Employee"/>
+    </owl:Class>
+
+    <owl:ObjectProperty rdf:about="http://example.org/ontology#worksFor">
+        <rdfs:label>works for</rdfs:label>
+        <rdfs:domain rdf:resource="http://example.org/ontology#Employee"/>
+        <rdfs:range rdf:resource="http://example.org/ontology#Organisation"/>
+    </owl:ObjectProperty>
+
+    <owl:ObjectProperty rdf:about="http://example.org/ontology#manages">
+        <rdfs:label>manages</rdfs:label>
+        <rdfs:domain rdf:resource="http://example.org/ontology#Manager"/>
+        <rdfs:range rdf:resource="http://example.org/ontology#Employee"/>
+        <owl:inverseOf rdf:resource="http://example.org/ontology#managedBy"/>
+    </owl:ObjectProperty>
+
+    <owl:ObjectProperty rdf:about="http://example.org/ontology#managedBy">
+        <rdfs:label>managed by</rdfs:label>
+        <rdfs:domain rdf:resource="http://example.org/ontology#Employee"/>
+        <rdfs:range rdf:resource="http://example.org/ontology#Manager"/>
+        <owl:inverseOf rdf:resource="http://example.org/ontology#manages"/>
+    </owl:ObjectProperty>
+
+    <owl:DatatypeProperty rdf:about="http://example.org/ontology#name">
+        <rdfs:label>name</rdfs:label>
+        <rdfs:domain rdf:resource="http://example.org/ontology#Person"/>
+        <rdfs:range rdf:resource="http://www.w3.org/2001/XMLSchema#string"/>
+    </owl:DatatypeProperty>
+
+    <owl:DatatypeProperty rdf:about="http://example.org/ontology#age">
+        <rdfs:label>age</rdfs:label>
+        <rdfs:domain rdf:resource="http://example.org/ontology#Person"/>
+        <rdfs:range rdf:resource="http://www.w3.org/2001/XMLSchema#integer"/>
+    </owl:DatatypeProperty>
+
+    <owl:DatatypeProperty rdf:about="http://example.org/ontology#email">
+        <rdfs:label>email</rdfs:label>
+        <rdfs:domain rdf:resource="http://example.org/ontology#Person"/>
+        <rdfs:range rdf:resource="http://www.w3.org/2001/XMLSchema#string"/>
+    </owl:DatatypeProperty>
+
+    <owl:DatatypeProperty rdf:about="http://example.org/ontology#foundedDate">
+        <rdfs:label>founded date</rdfs:label>
+        <rdfs:domain rdf:resource="http://example.org/ontology#Organisation"/>
+        <rdfs:range rdf:resource="http://www.w3.org/2001/XMLSchema#date"/>
+    </owl:DatatypeProperty>
+
+</rdf:RDF>
