+++ apps/desktop/src/renderer/src/App.tsx
@@ -47,6 +47,7 @@ import { useUIStore } from './store/ui';
 
 function App(): React.JSX.Element {
   const ontology = useOntologyStore((s) => s.ontology);
+  const loadFromFile = useOntologyStore((s) => s.loadFromFile);
   const loadFromTurtle = useOntologyStore((s) => s.loadFromTurtle);
   const exportToTurtle = useOntologyStore((s) => s.exportToTurtle);
   const setFilePath = useOntologyStore((s) => s.setFilePath);
@@ -72,9 +73,9 @@ function App(): React.JSX.Element {
   const handleOpen = useCallback(async () => {
     const result = await window.api.openFile();
     if (result) {
-      loadFromTurtle(result.content, result.filePath);
+      loadFromFile(result.content, result.filePath);
     }
-  }, [loadFromTurtle]);
+  }, [loadFromFile]);
 
   const doSave = useCallback(async () => {
     const turtle = exportToTurtle();
@@ -205,10 +206,10 @@ function App(): React.JSX.Element {
     async (filePath: string) => {
       const result = await window.api.openRecentFile(filePath);
       if (result) {
-        loadFromTurtle(result.content, result.filePath);
+        loadFromFile(result.content, result.filePath);
       }
     },
-    [loadFromTurtle],
+    [loadFromFile],
   );
 
   const hasContent = ontology.classes.size > 0;
